+++ OngekiFumenEditor/UI/ValueConverters/ParamValueToBoolConverter.cs
@@ -1,0 +1,18 @@
+using System;
+using System.Globalization;
+using System.Windows.Data;
+
+namespace OngekiFumenEditor.UI.ValueConverters;
+
+public class ParamValueToBoolConverter : IValueConverter
+{
+    public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        return value?.Equals(parameter);
+    }
+
+    public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+    {
+        return value?.Equals(true) == true ? parameter : Binding.DoNothing;
+    }
+}
